peluchegpt/backend/requirements.txt
# Dependências do backend PelucheGPT.
fastapi
uvicorn[standard]
httpx
pydantic
chromadb
sentence-transformers
google-generativeai
pyinstaller
